+++ src/components/RepairTaskCard/RepairTaskCard.jsx
@@ -3,7 +3,7 @@ import { useState } from 'react'
 import { Link } from 'react-router-dom'
 
 // components
-import EditRepairTaskCard from '../../EditRepairTaskCard/EditRepairTaskCard'
+import EditRepairTaskCard from '../EditRepairTaskCard/EditRepairTaskCard'
 
 // css
 // import styles from './RepairTaskCard.module.css'
@@ -32,7 +32,11 @@ const RepairTaskCard = ({ repairTask, repairId, user, handleDeleteRepairTask }) 
               </button>
             </>
           }
-          {/* {showEditRepairTask && <EditRepairTaskCard  /> } */}
+          {/* {showEditRepairTask && <EditRepairTaskCard  
+            repairId={repairId} 
+            user={user} 
+            repairTask={repairTask} 
+          />} */}
         </span>
       </header>
       <li>{repairTask.task} - {repairTask.done ? 'done' : 'not done' }</li>
